Moved randomness helper functions to their own header
